requirement.txt
Django==4.1.5
djangorestframework==3.14.0
psycopg2==2.9.5
django-extensions==3.2.1
Pillow==9.4.0
